Restructure src directory
Restructure the src directory in the code base to allow for easier
extensibility in future.
