+++ src/core/camera.h
