+++ src/core/colour.h
