+++ src/core/image.h
