+++ src/core/image_writer.h
