+++ src/core/ray.h
@@ -1,7 +1,7 @@
 #ifndef RAY_H
 #define RAY_H
 
-#include "vec3.h"
+#include "core/vec3.h"
 
 #include <stdexcept>
 #include <limits>
