+++ src/core/utilities.h
