+++ src/primatives/intersectable_list.h
@@ -1,8 +1,8 @@
 #ifndef INTERSECTABLE_LIST_H
 #define INTERSECTABLE_LIST_H
 
-#include "intersectable.h"
-#include "ray.h"
+#include "core/intersectable.h"
+#include "core/ray.h"
 
 #include <memory>
 #include <vector>
