+++ test/ray_test.cpp
@@ -1,5 +1,5 @@
-#include "ray.h"
-#include "vec3.h"
+#include "core/ray.h"
+#include "core/vec3.h"
 
 #include <gtest/gtest.h>
