+++ test/sphere_test.cpp
@@ -1,6 +1,6 @@
-#include <vec3.h>
-#include <sphere.h>
-#include <ray.h>
+#include "core/vec3.h"
+#include "primatives/sphere.h"
+#include "core/ray.h"
 #include <gtest/gtest.h>
 
 TEST(SphereTests, TestDeclaration) {
