+++ test/vec3_test.cpp
@@ -1,4 +1,4 @@
-#include <vec3.h>
+#include "core/vec3.h"
 #include <gtest/gtest.h>
 
 TEST(VectorTests, TestDefaultDeclaration) {
